train/requirements.txt
# To install these Python dependencies, please type:
# pip install -r requirements.txt
gensim==3.8.3
numpy==1.18.4
pandas==1.0.4
scikit-learn==0.23.1
sklearn==0.0
nltk==3.5
PyYAML==5.3.1
seaborn==0.10.1
